Improve PQconnectPoll by strictly parsing the keepalives parameter.
Commit e7a2217978 introduced the parse_int_param() function to parse
libpq connection parameters more strictly. However, PQconnectPoll()
was not updated to use parse_int_param() when parsing the keepalives
parameter. This caused inconsistency in the handling of the same
parameter between PQconnectPoll() and PQgetCancel().

This commit resolves the issue by using parse_int_param() to parse
the keepalives parameter in useKeepalives() called in PQconnectPoll().
By parsing the keepalives parameter more strictly, the behavior of
PQconnectPoll() is now consistent with PQgetCancel().