(WIP) Time-delayed logical replication by serializing changes
Similar to physical replication, a time-delayed copy of the data for
logical replication is useful for some scenarios (particularly to fix
errors that might cause data loss).

This patch implements a new subscription parameter called 'min_apply_delay'.

If the subscription sets min_apply_delay parameter, the logical
replication worker will delay the transaction apply for min_apply_delay
milliseconds.

The delaying is implemented by serializing changes into file. The file is
created when the worker receives BEGIN message. The worker writes received
changes and flush at COMMIT. The delayed transaction is checked its commit time
for every main loop, and applied from the file when the time exceeds the
min_apply_delay. The commit time is stored in memory when the transaction is
committed, or the worker restarts.

The delay is calculated between the WAL time stamp and the current time
on the subscriber.

The combination of parallel streaming mode and min_apply_delay is not
allowed. This is because in parallel streaming mode, we start applying
the transaction stream as soon as the first change arrives without
knowing the transaction's prepare/commit time. This means we cannot
calculate the underlying network/decoding lag between publisher and
subscriber, and so always waiting for the full 'min_apply_delay' period
might include unnecessary delay.

The other possibility was to apply the delay at the end of the parallel
apply transaction but that would cause issues related to resource
bloat and locks being held for a long time.

Earlier versions were written by Euler Taveira, Takamichi Osumi, and Kuroda Hayato

Author: Kuroda Hayato, Takamichi Osumi