Merge "basic vacuuming" sect2 into sect1 introduction.
This doesn't change any of the content itself.  It just merges the
original text into the sect1 text that immediately preceded it.

This is preparation for the next commit, which will remove most of the
text "relocated" in this commit.  This structure should make things a
little easier for doc translators.