Overhaul "Recovering Disk Space" vacuuming docs.
XXX This commit is much less worked out and polished than the work on
freezing.  It should very much be considered a work in progress, and
isn't the priority for now.

Say a lot more about the possible impact of long-running transactions on
VACUUM.  Remove all talk of administrators getting by without
autovacuum; at most administrators might want to schedule manual VACUUM
operations to supplement autovacuum (this documentation was written at a
time when the visibility map didn't exist, even in its most basic form).

Also describe VACUUM FULL as an entirely different kind of operation to
conventional lazy vacuum.

XXX Open question for this commit:

I wonder if it would make sense to move all of that stuff into its own
new sect1 of "Chapter 29. Monitoring Disk Usage" -- something along
the lines of "what to do about bloat when all else fails, when the
problem gets completely out of hand". Naturally we'd link to this new
section from "Routine Vacuuming".

XXX For now, a lot of the information about CLUSTER and VACUUM FULL is
moved into Note/Warning boxes.  This arrangement is definitely going to
be temporary.