Move Interpreting XID stamps from tuple headers.
Move handling of 32-bit XID comparisons/physical wraparound from
"Routine Vacuuming" to chapter about transaction internals.

This is intended to be fairly close to a mechanical change.  It isn't
entirely mechanical, though, since the original wording has been
slightly modified for it to work in context.

TODO fix xact.sgml indentation.  The new content is indented correctly
already, but the existing content will need to be re-indented to match
in a later commit.  As always, structuring things this way is intended
to make life a little bit easier for doc translators.