Make "Routine Vacuuming" autovacuum-orientated.
Now that it's no longer in its own sect2, shorten the "Vacuuming basics"
content, and make it more autovacuum-orientated.  This gives much less
prominence to VACUUM FULL, which has little place in a section about
autovacuum.  We no longer define avoiding the need to run VACUUM FULL as
the purpose of vacuuming.

A later commit that overhauls "Recovering Disk Space" will add back a
passing mention of things like VACUUM FULL and TRUNCATE, but only as
something that might be relevant in extreme cases.  (Use of these
commands is hopefully neither "Routine" nor "Basic" to most users).

Also add some introductory information about the audience and goals of
the "Routine Vacuuming" section of the docs.