Rename 'lpp' to 'lp' in heapam.c
In heapam.c we use a mixture of 'lpp'(13/125) and 'lp'(112/125) for the ItemId variable to iterate over the line pointer array. This is confusing as they represent the same thing. Previously there was code like '++lpp' and technically 'lpp' makes more sense because we are incrementing the line pointer's pointer(lpp) but not the line pointer(lp) itself.

Since 1. we have removed all the '++lpp' and now only increment lineoff during iteration and 2. there are 100+ more occurrences of 'lp' than 'lpp', I suggest we make the naming consistent by changing all the 'lpp' to 'lp' to avoid confusion.