Allow listening port to be 0
This makes it possible to start postgres on an unused port instead of
trying to reliably pick it in cases where this is desirable, such as
test benches or colocated database instances.

This is done by retrieving the assigned port number using `getsockname`.

Those using this feature can retrieve the assigned port through
postmaster.pid