Allow user to mark a transaction uncommittable.
In some cases, at the beginning of a transaction the user knows that the
transaction is not supposed to be committed. In such cases the user can
start the transaction with NOT COMMITTABLE attribute, and Postgres will
refuse to commit the transaction.

While a transaction is in progress, upon seeing an undesirable state,
the user may want to continue the transaction to troubleshoot or learn
more, but they may want to ensure that the transaction does not get
committed. For such cases, the user can use SET TRANSACTION NOT
COMMITTABLE command, and Postgres will refuse to commit the transaction.

The user may use SET SESSION CHARACTERISTICS commmand to change the
default COMMITTABLE attribute of all future transactions.

Once marked as NOT COMMITTABLE, a transaction cannot then be marked as
COMMITTABLE.