Storage for session variables and SQL interface
Session variables are stored in session memory in dedicated hash table. The
write access is done by LET command. The read access is by SELECT command.
The access rights should be checked.

The identifiers of session variables should be shadowed by possible column's
identifiers always. This is by design feature. We don't want to break an
application by creating some session variable that is badly named.

The limits of this patch (solved by other patches):

- session variables block parallel execution
- session variables blocks simple expression evaluation (in plpgsql)
- SQL functions with session variables are not inlined
- CALL statement is not supported (usage of direct access to express executor)
- memory used by dropped session variables is not released

Implementations of EXPLAIN LET and PREPARE LET statements
are in separate patches (for better readability)