Support prepared statement invalidation when result types change
The cached plan for a prepared statements can get invalidated when DDL
changes the tables used in the query, or when search_path changes. When
this happens the prepared statement can still be executed, but it will
be replanned in the new context. This means that the prepared statement
will do something different e.g. in case of search_path changes it will
select data from a completely different table. This won't throw an
error, because it is considered the responsibility of the operator and
query writers that the query will still do the intended thing.

However, we would throw an error if the result of the query is of a
different type than it was before. This was not documented anywhere and
can thus be a surprising error to hit. It doesn't need to be an error
as long as we send the correct RowDescription. Then clients can use that
to handle the case when the result types or column counts change in
between EXECUTE calls. Most clients already do this anyway, and for any
that don't it should not be a big issue to implement it, since it
requires no extra round-trip and it's already needed for normal queries
anyway.

This patch allows a prepared statement to continue to work even when the
result type changes.

Without this change all clients that automatically prepare queries as a
performance optimization will need to handle or avoid the error somehow,
often resulting in deallocating and re-preparing queries when its
usually not necessary. With this change connection poolers can also
safely prepare the same query only once on a connection and share this
one prepared query across clients that prepared that exact same query.