Allow any constant types in GROUP BY clause
In v14 and earlier, only integer Consts were allowed in the GROUP BY
clause.  Boolean constants made it through as these had no native node
type and were expressed as TypeCasts.  As of 941460fcf, Booleans have
their own node type which resulted in "GROUP BY true" no longer working in
v15 and v16.

Here adjust the code to remove the special Boolean case that was added
to fix the v15/v16 regression and just allow any constant types in the
GROUP BY.  Integers, of course, remain a special case to reference the
ordinal position of the column in the target list to group by.

Also, clarify in the documentation that references made in the GROUP BY
to output columns must be done so using integer literals.  This should
lessen the chances of any surprises that "GROUP BY 1.0" doesn't group by
the 1st column in the target list.

Discussion: https://postgr.es/m/CAApHDvrSUvb1ODvNcYGHz4O6WEsPChhwmSkJcs_3y5pniN2p+A@mail.gmail.com