Document the oidvector type
The oidvector type is used in pg_proc.proargtypes, and perhaps
elsewhere in the catalogs.  But there is no documentation on how to
get oids out of a oidvector or otherwise manipulate the data type for
those who wish to do so.  This seems the only place where such
documentation would go.