Introduce initdb-selectable block sizes
Whereas we had traditionally used BLCKSZ to indicate the compile-time selected
block size, this commit adjusted things so the cluster block size can be
selected at initdb time.

In order to code for this, we introduce a few new defines:

- cluster_block_size is the blocksize for this cluster itself.  This is not
valid until BlockSizeInit() has been called in the given backend, which we do as
early as possible by parsing the ControlFile and using the blcksz field.

- MIN_BLOCK_SIZE and MAX_BLOCK_SIZE are the limits for the selectable block
size. It is required that cluster_block_size is a power of 2 between these two
constants.

- DEFAULT_BLOCK_SIZE is the moral equivalent of BLCKSZ; it is the built-in
default value.  This is used in a few places that just needed a buffer of an
arbitrary size, but the dynamic value cluster_block_size should almost always be
used instead.

- cluster_relseg_size is used instead of RELSEG_SIZE, since internally we are
storing the segment size in terms of number of blocks.  RELSEG_SIZE is still
kept, but is used in terms of the number of blocks of DEFAULT_BLOCK_SIZE;
cluster_relseg_size scales appropriately (and is the only thing used internally)
to keep the same target total segment size regardless of block size.

This patch uses inline functions to switch between compile-time constants for
all blocksize-derived fields; this makes what could be a heavy computation or
memory load/store in hot code paths a relatively fast (~ 4 instructions)
computation for any access to such a dynamic value. Where possible, the
traditional macro names are kept so code that doesn't care about it should not
need to change, however the definition of these has changed (see the CalcXXX()
routines in blocksize.h for details).

A new function, BlockSizeInit() populates the appropriate values based on the
target block size.  This should be called as early as possible in any code that
utilizes block sizes.  This patch adds this in the appropriate place on the
handful of src/bin/ programs that used cluster_block_size, so this caveat mainly impacts new
code.

Headers which define now dynamic values should declare the symbol name and the
calculation macro in order to properly register it; BlockSizeDecl() ends up
creating the necessary static inline function, but only in files which include
the given headers.

Code which had previously used BLCKZ should likely be able to get away with
changing these instances to cluster_block_size, unless you're using a structure
allocated on the stack.  In these cases, the compiler will complain about
dynamic structure.  The solution is to utilize an expression with MAX_BLOCK_SIZE
instead of cluster_block_size, ensuring enough stack space is allocated for the maximum
size. This also does require using cluster_block_size or an expression based on
it when actually using this structure, so in practice more stack space may be
allocated then used in principal; as long as there is plenty of stack this
should have no specific impacts on code.

From what I can tell, TPC-H results are in the neighborhood of the equivalent
unpatched version, which is promising compared to the previous version.