optimization: Add support for fast, non-division-based div/mod algorithms
One place that is unhappy with the runtime changes is the visibility map.  This
provides a feature for a fast mod/div operation using only a single cacheable
division operation to compute the inverse which we can then multiply and
bitshift appropriately for the actual computations, which serves our needs here.