Add encrypted/authenticated WAL
When using an encrypted cluster, we need to ensure that the WAL is also
encrypted. While we could go with an page-based approach, we use instead a
per-record approach, using GCM for the encryption method and storing the AuthTag
in the xl_crc field.

We change the xl_crc field to instead be a union struct, with a compile-time
adjustable size to allow us to customize the number of bytes allocated to the
GCM authtag.  This allows us to easily adjust the size of bytes needed to
support our authentication.  (Testing has included up to 12, but leaving at this
point to 4 due to keeping the size of the WAL records the same.)