Skip Orderby clause execution for Materialized Views
The order is not guaranteed even if materialized view's sql has
order by clause.

Materialized view’s order by clause could be skipped when execution
if order by clause is on the top query level and there is no real
limit clause.

This will speed up for REFRESH and CREATE AS on materialized views.

Authored-by: Zhang Mingli avamingli@gmail.com