Predicate locks are held per-cluster, not per-database
I believe this is still true and nothing has changed since
this email thread:
https://www.postgresql.org/message-id/20170518110712.31caaf3b@slate.meme.com

This is a significant corner case and so should be documented.  It is
also somewhat suprising since the databases within a cluster are
otherwise isolated, at least from the user's perspective.

I mention psql pagination because it is easy to overlook the fact that
it is not just end-user applications that generate transactions.

I have not thought about this in a long time and am not 100% certain
that WAL checkpoints on segments are delayed until the predicate lock
for the respective transaction is released.  I'm pretty sure, but this
should be checked.

I am sure that all concurrent predicate locks must be released
together.  See the 2nd to last sentence of ports12:
https://arxiv.org/pdf/1208.4179.pdf

The index entries might be improved.  The idea is to guide someone
who's got a lot of WAL laying around and is using serializable
transactions.