Add null termination to string received in parallel apply worker.
The parallel apply worker didn't add null termination to the string received
from the leader apply worker via the shared memory queue. This action violates
the rule established in StringInfo, which guarantees the presence of a
terminating '\0' at the end of the string.

So fix it by using standard StringInfo to store the received string.