pg_upgrade: use upgrade function to restore OID
Previously, the OID counter is restored by invoking pg_resetwal with the -o
option, at the end of upgrade. This is not problematic for now, but WAL removals
are not happy if we want to upgrade logical replication slot.

Based on above, this commit adds a new upgrade function to reset next OID.
SetNextObjectId() has a ratchet mechanism to avoid a rollback of OIDs, but it is
ignored during the upgrade: wraparound of the OID counter might occur on the old
cluster, but pg_upgrade tries to restore the counter to the new cluster anyway.

Also, an obsolated comment was removed. Controlfile will not be updated so that
we do not have to take care anything for primary server.

Note that we only update the on-memory data to avoid concurrent update of
control with the chekpointer. It is harmless because the value would be set at
shutdown checkpoint.