Use PGAlignedBlock instead of "char buf[BLCKSZ]" in more places
Commit 44cac934 replaced "char buf[BLCKSZ]" with PGAlignedBlock to
avoid issues on alignment-picky hardware. While it replaced most
of the instances, there are still some more left. This commit
uses PGAlignedBlock in the left-over places. A note in the commit
44cac934 says that ensuring proper alignment makes kernel data
transfers fasters and the left-over "char buf[BLCKSZ]" either do
read() or write() system calls, so it might be worth to align them
with PGAlignedBlock.