Fix segfault when notifying in a ProcessUtility hook
This can happen when using a transaction block and a ProcessUtility hook
that would send a notification on Rollback.

When a transaction block fails, it will enter in a TBLOCK_ABORT state,
waiting for a rollback. Calling rollback will switch to a
TBLOCK_ABORT_END state and will only go through CleanupTransaction.
If a hook has sent a notification during the rollback command, a
notification will be queued but its content will be wiped during memory
cleanup.
Trying to send a notification immediately after will segfault in
PreCommit_Notify as pendingNotifies->events will be invalid.

Fix by moving cleanup notification from AbortTransaction to
CleanupTransaction.