Add IO stream abstraction
This is a shared abstraction between the frontend and backend that
is designed to draw a clean line between the compoments of postgres
that want to consume "regular" protocol bytes and any layers that
may want to transform them. In this patch, encyryption is simply
moved to use this abstraction rather than callers directly using
secure_read/secure_write (which depending on context might or might
not actually be secure). This is a pre-work change intended to make
compression integrate into frontend and backend comms layers more
cleanly.

This may be a first step towards facilitating greater sharing of
SSL/GSS api code between the frontend and backend down the road,
but for now it the stream processors themselves are completely
independent.