Fix regression tests caused by additional Result nodes on top of Sorts et al.
We not add an extra Result node to project away the unneeded ORDER BY
junk columns. This may seem bad from a performance point of view, but
remember that previously the junk filter in the executor was doing
basically the same work anyway. It's just represented in the plan tree
now.

Perhaps we should teach Sort and MergeAppend nodes how to project to
avoid this churn.

(This commit should be squashed with the previous commit that caused
these diffs, but I'm keeping them separate for now to make it easier
to see the effect of the previous commit.)