postgres_fdw: child-join with ConvertRowtypeExprs causes expression reference errors.
Whole row reference of the parent is translated into
ConvertRowtypeExpr with whole row of child as an argument. When
partition-wise join is used, targetlist of a child-join contains
ConvertRowtypeExpr/s when the parent-join's targetlist has whole-row
reference/s of joining partitioned tables.

The targetlist to be deparsed for a child-join is built using
build_tlist_to_deparse(). The same targetlist is then saved as
fdw_scan_tlist in ForeignScanPlan. build_tlist_to_deparse() pulls only
Var nodes from the join's targetlist.  So it pulls Var reprensenting a
whole-row reference of a child from a ConvertRowtypeExpr. Thus the
child-join's projected target list contains ConvertRowtypeExpr but the
SELECT clause and fdw_scan_tlist contains a whole-row reference.

This causes two problems:
1.  When a relation (participating in the child-join being pushed
down) is deparsed as a subquery, subquery's targetlist (and SELECT
clause) contains expressions from foreignrel->reltarget->exprs. A Var
from such a relation is deparsed as column reference of subquery using
get_relation_column_alias_ids().  That function uses foreignrel's
targetlist to locate given node to be deparsed.  If the joining
relation corresponding to ConvertRowtypeExpr is deparsed as a
subquery, this function is called with whole-row reference node (Var
node with varattno = 0).  The relation's foreignrel->reltarget->exprs
doesn't contain its whole-row reference directly but has it embedded
in ConvertRowtypeExpr. So, the function doesn't find the given node
and throws error.

2. When there is possibility of EvalPlanQual being called, we
construct local join plan matching the pushed down foreign join. In
postgresGetForeignPlan() after we have built the local join plan, the
topmost plan node's targetlist is changed to fdw_scan_tlist to match
the output of the ForeignScan node. As explained above, this
fdw_scan_tlist contains a bare reference to the whole-row reference from a
child relation if the child-join's targetlist contains a
ConvertRowtypeExpr. When changing the topmost plan node's targetlist,
we do not modify the targetlists of its left and right tree nodes. The
left/right plan involving corresponding child relation will have
ConvertRowtypeExpr expression in its targetlist, but not whole-row
reference directly. When the topmost local join plan node's targetlist
is processed by set_plan_ref(), it throws error "variable not found in
subplan target lists" since it doesn't find bare whole-row reference
of the child relation in subplan's targetlists.

Solution
This requires two parts
a. In build_tlist_to_deparse(), instead of pulling Var node from
ConvertRowtypeExpr, we pull whole ConvertRowtypeExpr and include it in
the targetlist being deparsed which is also used to set
fdw_scan_tlist.

b. deparse ConvertRowtypeExpr.  For this we need to get the conversion
map between the parent and child. We then deparse ConvertRowtypeExpr
as a ROW() with the attributes of child rearranged per the conversion
map. A multi-level partitioned table will have nested
ConvertRowtypeExpr. To deparse such expressions, we need to find the
conversion map between the topmost parent and the child, by ignoring
any intermediate parents.

The patch is taken from https://www.postgresql.org/message-id/CAFjFpRc8ZoDm0%2Bzhx%2BMckwGyEqkOzWcpVqbvjaxwdGarZSNrmA%40mail.gmail.com