feature: Add hook for setting reloptions defaults at runtime
This is necessary due to the various toast constants being non-constant due to
runtime differences in usable space.  Future commits will utilize this to tweak.

This is run once on startup, so while this is not the most efficient way to
handle this, it doesn't really matter for one-time spinup cost.