Track transactions committed in BUILDING_SNAPSHOT.
We previously didn't track transactions that begin after BUILDING_SNAPSHOT
and commit before FULL_SNAPSHOT when building historic snapshot in logical
decoding. This can cause transactions that begin after FULL_SNAPSHOT to take
an incorrect snapshot because transactions committed in BUILDING_SNAPSHOT
are not in the snapshot.

To fix it, we track transactions that committed in BUILDING_SNAPSHOT and
add them to historic snapshot if they have catalog changes. Note that
there is no need to track transactions that begin before BUILDING_SNAPSHOT
because they are all finished when we reach FULL_SNAPSHOT.