btree: Introduce basic specialization
This introduces the basic infrastructure for specialization, and updates
call sites of the specialized functions to use the new infrastructure.

Note that actual meaningful code behaviour changes will happen in a later
commit, this only builds the scaffolds.

Specialized functions are written in _spec files, and their names are
defined as macros that produce either a specialized name token (when in a
specialized context), or an expression (which utilizes an in-scope
nbts_prep_ctx-prepared context) that evaluates to a function pointer which
points to the indicated function's variant which is most optimized for that
context's index key shape.