Improve pg_trigger.tgargs representation
Before, the trigger argument list was encoded into a bytea column.
Replace this with a list of nodes in a pg_node_tree.  This makes the
internal code much simpler and also allows clients to decode the
column more easily.

Clients shipped with PostgreSQL (such as psql and pg_dump) use
pg_get_triggerdef() and don't access the field directly, so they need
to no changes.

TODO: catversion