Update to_regtype docs regarding error condition
The statement that `to_regtype` returns `NULL` rather than raising an
error turns out to be false when the parser cannot extract a type. This
can happen for inputs such as `inteval second`, where the SQL grammar
parser fails to parse the type name. In most cases it does return `NULL`
and ideally it always would for such unknown types, but sadly the
overhead of handling errors thrown by the grammar is too high at this
point. So document the variation, instead.