pgcrypto: Remove hash speed comparison table
The table comparing relative hash speed is at best misleading, and
at worst misinforming.  The benchmarks are old and do not reflect
the current state of the art.  The table also imply security due to
brute-force protection, which is no longer true for many algorithms
which have published vulnerabilities which avoid the need for brute
force.